requirements.txt
absl-py @ file:///tmp/build/80754af9/absl-py_1607439979954/work
aiohttp @ file:///C:/ci/aiohttp_1602530153430/work
astunparse==1.6.3
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
blinker==1.4
blis @ file:///C:/ci/cython-blis_1594248803162/work
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1607706694405/work
catalogue==1.0.0
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1606255208697/work
chardet @ file:///C:/ci/chardet_1605303225733/work
click @ file:///home/linux1/recipes/ci/click_1610990599742/work
cryptography @ file:///C:/ci/cryptography_1607637849569/work
cymem @ file:///C:/ci/cymem_1605062370412/work
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-2.3.1/en_core_web_md-2.3.1.tar.gz
Flask==1.1.2
gast @ file:///tmp/build/80754af9/gast_1597433534803/work
google-auth @ file:///tmp/build/80754af9/google-auth_1607969906642/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1603929124518/work
google-pasta==0.2.0
grpcio @ file:///C:/ci/grpcio_1597406462198/work
gunicorn==20.0.4
h5py==2.10.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
itsdangerous==1.1.0
Jinja2 @ file:///home/linux1/recipes/ci/jinja2_1610990516718/work
jsonschema @ file:///C:/ci/jsonschema_1594295134295/work
Keras @ file:///tmp/build/80754af9/keras_split_1593112142734/work
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing==1.1.0
Markdown @ file:///C:/ci/markdown_1605111189761/work
MarkupSafe==1.1.1
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict @ file:///C:/ci/multidict_1600456481656/work
murmurhash @ file:///C:/ci/murmurhash_1607450221995/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1603466732592/work
oauthlib==3.1.0
opt-einsum==3.1.0
pandas @ file:///C:/ci/pandas_1611171485393/work
plac @ file:///C:/ci/plac_1594260075734/work
preshed==3.0.2
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyJWT==1.7.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyreadline==2.1
pyrsistent @ file:///C:/ci/pyrsistent_1600141795814/work
PySocks @ file:///C:/ci/pysocks_1605287845585/work
python-dateutil==2.8.1
python-decouple==3.4
pytz @ file:///tmp/build/80754af9/pytz_1608922264688/work
PyYAML==5.4
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1610483308194/work
scipy @ file:///C:/ci/scipy_1604596260408/work
six @ file:///C:/ci/six_1605187374963/work
spacy @ file:///C:/ci/spacy_1608303051812/work
srsly @ file:///C:/ci/srsly_1607542026061/work
tensorboard @ file:///home/builder/ktietz/conda/conda-bld/tensorboard_1604313476433/work/tmp_pip_dir
tensorboard-plugin-wit==1.6.0
tensorflow==2.3.0
tensorflow-estimator @ file:///tmp/build/80754af9/tensorflow-estimator_1599136169057/work/whl_temp/tensorflow_estimator-2.3.0-py2.py3-none-any.whl
termcolor==1.1.0
thinc @ file:///C:/ci/thinc_1607704777175/work
tqdm @ file:///tmp/build/80754af9/tqdm_1609788246169/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1606938623459/work
wasabi @ file:///tmp/build/80754af9/wasabi_1599244993840/work
Werkzeug==1.0.1
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wincertstore==0.2
wrapt==1.12.1
yarl @ file:///C:/ci/yarl_1606940076464/work
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
